Implement ravel and unravel for sparseCOO matrices

+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ndarray/BaseNDArray.java
@@ -6532,7 +6532,7 @@ public abstract class BaseNDArray implements INDArray, Iterable {
     }
 
     @Override
-    public int[] sparseOffsets() {
+    public long[] sparseOffsets() {
         throw new UnsupportedOperationException("Not a sparse ndarray");
     }
