+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ndarray/INDArray.java
@@ -3033,7 +3033,7 @@ public interface INDArray extends Serializable {
 
     int[] hiddenDimensions();
 
-    int[] sparseOffsets();
+    long[] sparseOffsets();
 
     int underlyingRank();
